requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pymysql
psycopg2-binary
cryptography
alembic
pydantic
httpx
gunicorn